Attempt to Parallelize Image Updates
Problem is that _asynchrnous_ update mechanism depends on color to be the on the canvas when an update is dispatched. When a new update COMPLETES between another call and COMPLETE, the update mechanism fails.

Next attempt: store “shadow” copy of sigmoid result in state (on component load), on props change call a WW to apply colors to sigmoid state.